+++ libbulletml/src/formula-variables.cpp
@@ -1,7 +1,7 @@
 #include "formula-variables.h"
 
 namespace Variables {
-	double rank;
-	std::vector<double>* parameters;
+	float rank;
+	std::vector<float>* parameters;
 	BulletMLRunner* runner;
 }
